+++ v2/cmd/integration-test/http.go
@@ -165,7 +165,7 @@ func (h *httpInteractshStopAtFirstMatchRequest) Execute(filePath string) error {
 		return err
 	}
 	// polling is asyncronous, so the interactions may be retrieved after the first request
-	return expectResultsCount(results, 1, 2)
+	return expectResultsCount(results, 1)
 }
 
 type httpGetHeaders struct{}
